refactor: refactoring the sign in page into components
